requirements.txt
enum34==1.0
httplib2==0.9
httpretty==0.8.0
mock==1.0.1
oauth2client==1.2
python-dateutil>=2.2
requests==2.3.0
six==1.6.1
sure==1.2.7
nose
